Added revoke funtion and its testcase
As more ExecuteMsg are added to the contract, msg.rs and error.rs are
updated to suit the need of the custom contract. In Cw721, the token_id
is String, whereas here, the token_id is u64. This is why the files written
earlier to match the 721 spec are requiring changes.